Move helper method to its only usage
It makes `selector_for` easier to understand.
