--- features/support/selectors.rb
@@ -1,44 +1,0 @@
-module HtmlSelectorsHelpers
-  # Maps a name to a selector. Used primarily by the
-  #
-  #   When /^(.+) within (.+)$/ do |step, scope|
-  #
-  # step definitions in web_steps.rb
-  #
-  def selector_for(locator)
-    case locator
-
-    # Add more mappings here.
-    # Here is an example that pulls values out of the Regexp:
-    #
-    #  when /^the (notice|error|info) flash$/
-    #    ".flash.#{$1}"
-
-    # You can also return an array to use a different selector
-    # type, like:
-    #
-    #  when /the header/
-    #    [:xpath, "//header"]
-
-    when "index grid"
-      [:css, "table.index_grid"]
-
-    when /^the "([^"]*)" sidebar$/
-      [:css, "##{$1.tr(" ", '').underscore}_sidebar_section"]
-
-    # This allows you to provide a quoted selector as the scope
-    # for "within" steps as was previously the default for the
-    # web steps:
-    when /^"(.+)"$/
-      $1
-
-    else
-      # :nocov:
-      raise "Can't find mapping from \"#{locator}\" to a selector.\n" +
-        "Now, go and add a mapping in #{__FILE__}"
-      # :nocov:
-    end
-  end
-end
-
-World(HtmlSelectorsHelpers)
